/watchesのtestを/current_user/watchesのtestへ転記し修正
